Add Descope auth to Remote Server
