+++ package.json
@@ -10,7 +10,7 @@
     "start:stdio": "node build/index.js",
     "start:sse": "node build/sse.js",
     "start": "node build/index.js",
-    "dev": "tsc --watch & nodemon build/index.js"
+    "dev": "tsc --watch & nodemon build/sse.js"
   },
   "files": [
     "build"
@@ -22,6 +22,7 @@
     "@descope/node-sdk": "^1.6.13",
     "@modelcontextprotocol/sdk": "^1.4.1",
     "express": "^4.21.2",
+    "http-proxy-middleware": "^3.0.3",
     "node-fetch": "^3.3.2",
     "zod": "^3.24.1"
   },
