+++ src/index.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
-import { createServer } from "./descope.js";
+import { createServer } from "./create-server.js";
 
 async function main() {
     const transport = new StdioServerTransport();
